Changed it back to old version. Version 105 does not exist anymore at this point. Could not show that a significant amount of RAM was saved.

+++ RealPhy_merge/src/buildTreesFromSequenceData/GetPolymorphismsClass.java
@@ -382,7 +382,7 @@ public abstract class GetPolymorphismsClass implements GetPolymorphisms,Serializ
 					queryIDColumn.add(queryID);
 				}
 				if(printInvariant||addition){
-					columns.addColumn(queryIDColumn, baseColumn,(byte)ref);
+					columns.addColumn(queryIDColumn, baseColumn,ref);
 				}
 			}
 		}
